task_1_modules/package_task1/requirements.txt
ansicolors    1.1.8
argcomplete   2.0.0
pip           21.3.1
rlane-libcli  1.0.3
rlane-netflix 1.0.2
setuptools    60.2.0
tomli         2.0.1
wheel         0.37.1
datetime